requirements.txt
####### example-requirements.txt #######
#
###### Requirements without Version Specifiers ######
pandas
streamlit
numpy
Pillow
